+++ src/main/java/com/estsoft/findmember_team01/information/controller/InformationViewController.java
@@ -7,10 +7,9 @@ import com.estsoft.findmember_team01.information.dto.InformationRequest;
 import com.estsoft.findmember_team01.information.dto.InformationView;
 import com.estsoft.findmember_team01.information.service.CommentService;
 import com.estsoft.findmember_team01.information.service.InformationService;
-import com.estsoft.findmember_team01.login.domain.Member;
+import com.estsoft.findmember_team01.member.domain.Member;
 import java.util.List;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
