+++ src/main/java/com/estsoft/findmember_team01/information/domain/Comment.java
@@ -2,7 +2,7 @@ package com.estsoft.findmember_team01.information.domain;
 
 
 import com.estsoft.findmember_team01.information.dto.CommentResponse;
-import com.estsoft.findmember_team01.login.domain.Member;
+import com.estsoft.findmember_team01.member.domain.Member;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
